local-cluster: extract faucet yaml-file writing functionality as a separate module.
